--- SECURITY.md
@@ -1,6 +1,0 @@
-# Security Policy
-
-## Reporting a Vulnerability
-
-Please email bugbounty@databricks.com to report any security vulnerabilities. We will acknowledge receipt of your vulnerability and strive to send you regular updates about our progress. If you're curious about the status of your disclosure please feel free to email us again. If you want to encrypt your disclosure email, you can use [this PGP key](https://keybase.io/arikfr/key.asc).
-
