+++ config.json
@@ -1,0 +1,12 @@
+{
+  "lakehouse": {
+    "catalog": "db_omics_solac",
+    "schema": "tcga_dev",
+    "volume": "tcga_files"
+  },
+  "api_paths": {
+    "cases_endpt": "https://api.gdc.cancer.gov/cases",
+    "files_endpt": "https://api.gdc.cancer.gov/files",
+    "data_endpt": "https://api.gdc.cancer.gov/data/"
+  }
+}
